Create category bottom sheet details and edit and remove

+++ app/src/main/java/com/example/budgetmanagement/database/Rooms/CategoryRepository.java
@@ -56,6 +56,9 @@ public class CategoryRepository {
     }
 
     public void delete(Category category) {
-        BudgetRoomDatabase.databaseWriteExecutor.execute(() -> categoryDao.delete(category));
+        int mandatoryCategoryId = 1;
+        if (category.getCategoryId() != mandatoryCategoryId) {
+            BudgetRoomDatabase.databaseWriteExecutor.execute(() -> categoryDao.delete(category));
+        }
     }
 }
